+++ ignite/services/chain/init.go
@@ -139,7 +139,14 @@ func (c *Chain) InitAccounts(ctx context.Context, cfg *chainconfig.Config) error
 
 		// If the account doesn't provide an address, we create one
 		if accountAddress == "" {
-			generatedAccount, err = commands.AddAccount(ctx, account.Name, account.Mnemonic, account.CoinType)
+			generatedAccount, err = commands.AddAccount(
+				ctx,
+				account.Name,
+				account.Mnemonic,
+				account.CoinType,
+				account.AccountNumber,
+				account.AddressIndex,
+			)
 			if err != nil {
 				return err
 			}